vptstream fixes
The vptstream component was not working because of inconsistent reference to the "threejs" lib in the depthkit module. 

The depthkit module has been updated, shrunk to the bare minimum so it's is not including three, hls in it's build. 

In doing this we had to hack in a global reference to HLS as it is not being included in dephkitjs anymore.